% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CompPepSetTestWorkflow.R
\name{CompPepSetTestWorkflow}
\alias{CompPepSetTestWorkflow}
\title{Competitive Peptide Set Test Workflow}
\usage{
CompPepSetTestWorkflow(
  dat,
  contrasts.par,
  group,
  pep_mapping_tbl,
  stat = c("t", "logFC"),
  correlated = FALSE,
  equal.correlation = FALSE,
  pepC.estim = c("sd", "mad"),
  logged = FALSE
)
}
\arguments{
\item{dat}{a dataframe or matrix of peptide abundance, or a SummarizedExperiment object where
grouping and peptide-protein mapping are provided in colData and rowData, respectively.}

\item{contrasts.par}{group levels to be compared separated by dash (e.g., "B-A"
if group B is to be compared against group A)}

\item{group}{a vector of group levels corresponding to each sample. Alternatively, it can be the
column name of the group in colData if dat is a SummarizedExperiment object.}

\item{pep_mapping_tbl}{a table mapping peptides to proteins. Alternatively, it can be the
column name of the protein in rowData if dat is a SummarizedExperiment object.}

\item{stat}{statistics to be used in the peptide set test.
Options include "t" (t-statistic) and "logFC" (log2 fold change).}

\item{correlated}{Boolean variable indicating whether peptides are assumed to be correlated.
If correlated, inter-peptide correlation will be estimated.}

\item{equal.correlation}{Boolean variable indicating whether all pairwise
inter-peptide correlation coefficients are assumed to be equal within a protein.
If true, the mixed model approach will be applied; otherwise, the approach
described in Wu and Smyth (2012), \emph{Nucleic Acids Research} will be applied. Note
that this parameter matters only if "correlated" is set to true.}

\item{pepC.estim}{estimator of the variance of peptide-wise t-statistics not
belonging to the protein of interest, i.e., test set. Options include "sd" and "mad".
"sd" represents sample standard deviation. "mad" represents sample median
absolute deviation.}

\item{logged}{Boolean variable indicating whether abundance data have been
log-transformed}
}
\value{
\code{CompPepSetTestWorkflow} returns a dataframe containing the following columns
\item{protein}{unique protein identifier}
\item{NPeps}{number of peptides}
\item{Correlation}{inter-peptide correlation coefficient}
\item{Direction}{direction of change}
\item{PValue}{raw p-value}
\item{adj.P.Val}{p-value adjusted via the Benjamini-Hochberg method}
\item{logFC}{average log2 fold change of peptides}
\item{Up}{number of upregulated peptides}
\item{Down}{number of downregulated peptides}
}
\description{
Given peptide abundance and assignment of peptide sequences to proteins,
execute the competitive peptide set test workflow to compute the log2 fold change,
p-value, and adjusted p-value of all proteins identified.
}
\examples{
# Generate random peptide data
dat <- 2^matrix(rnorm(3000), ncol = 6)
colnames(dat) <- paste0("Sample", 1:6)
rownames(dat) <- paste0("Peptide", 1:500)

# Generate peptide mapping table
pep_mapping_tbl <- data.frame(peptide = paste0("Peptide", 1:500),
protein = paste0("Protein", rep(1:100, each = 5)))

# Generate groups and contrasts
group <- c(rep("A", 3), rep("B", 3))
contrasts.par <- "B-A"

CompPepSetTestWorkflow(dat, contrasts.par = contrasts.par,
group = group,
pep_mapping_tbl = pep_mapping_tbl,
stat = "t",
correlated = TRUE,
equal.correlation = TRUE,
pepC.estim = "mad",
logged = FALSE)

# Store data as a SummarizedExperiment object
library(tibble)
library(SummarizedExperiment)
colData <- data.frame(sample = LETTERS[seq_along(group)], group = group) |> 
column_to_rownames(var = "sample")
rowData <- pep_mapping_tbl |> column_to_rownames(var = "peptide")
dat.nn <- dat
rownames(dat.nn) <- NULL
colnames(dat.nn) <- NULL
dat.se <- SummarizedExperiment(assays = list(int = dat.nn), colData = colData, rowData = rowData)

CompPepSetTestWorkflow(dat.se, contrasts.par = contrasts.par,
group = "group",
pep_mapping_tbl = "protein",
stat = "t",
correlated = TRUE,
equal.correlation = TRUE,
pepC.estim = "mad",
logged = FALSE)

}
\references{
Wu, D, and Smyth, GK (2012). Camera: a competitive gene set test accounting for inter-gene correlation. \emph{Nucleic Acids Research} 40, e133
}
\author{
Junmin Wang
}
